requirements.txt
discord.py
asyncio
random
